+++ src/trx/game/output/bind.c
@@ -3,6 +3,7 @@
 #include <trx/game/items.h>
 #include <trx/game/rooms/common.h>
 #include <trx/game/rooms/const.h>
+#include <trx/game/viewport.h>
 
 #include <string.h>
 
@@ -24,6 +25,10 @@ void Output_Bind_ResetRooms(void)
     for (int32_t i = 0; i < MAX_ROOMS; i++) {
         m_RoomBindings[i].active = false;
         m_RoomBindings[i].drawn = false;
+        m_RoomBindings[i].bound_left = Viewport_GetMaxX(VIEWPORT_GAME);
+        m_RoomBindings[i].bound_top = Viewport_GetMaxY(VIEWPORT_GAME);
+        m_RoomBindings[i].bound_right = Viewport_GetMinX(VIEWPORT_GAME);
+        m_RoomBindings[i].bound_bottom = Viewport_GetMinY(VIEWPORT_GAME);
     }
 }